perf: recent chart seperate file
